// src/models/Customer.ts
export type CustomerRow = {
  id: string;
  name: string;
  phone: string;
  email: string;
  password: string;
  company_name: string;
  location: string;
  created_at: string;
};

export type CustomerProfile = {
  id: string;
  name: string;
  phone: string;
  email: string;
  companyName: string;
  location: string;
  createdAt: string;
};